pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sdev"
version = "0.1.3"
description = "串口控制器工具包"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "klrc", email = "144069824@qq.com"},
]
keywords = ["serial", "controller", "hardware", "embedded"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Hardware",
]
requires-python = ">=3.7"
dependencies = [
    "pyserial>=3.5",
    "loguru>=0.6.0",
]

[project.urls]
Homepage = "https://github.com/klrc/sdev"
Repository = "https://github.com/klrc/sdev"
"Bug Tracker" = "https://github.com/klrc/sdev/issues" 